requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
beautifulsoup4==4.12.2
requests==2.31.0
language-tool-python==2.7.1
pyspellchecker==0.8.1
python-dotenv==1.0.0

